+++ todo-mvc-kitchensink/src/todoProcesses.ts
@@ -1,7 +1,7 @@
 import { createCommandFactory, createProcess } from '@dojo/framework/stores/process';
 import { PatchOperation } from '@dojo/framework/stores/state/Patch';
 import { add, remove, replace } from '@dojo/framework/stores/state/operations';
-import uuid from '@dojo/framework/core/uuid';
+import { uuid } from '@dojo/framework/core/util';
 
 export type TodoStore = {
 	completedCount: number;
