+++ todo-mvc-kitchensink/src/widgets/TodoApp.ts
@@ -3,14 +3,16 @@ import { v, w } from '@dojo/framework/widget-core/d';
 import ThemedMixin, { theme } from '@dojo/framework/widget-core/mixins/Themed';
 
 import ThemeSwitcherContainer from './../containers/ThemeSwitcherContainer';
-import TodoListOutlet from './../outlets/TodoListOutlet';
-import TodoFooterOutlet from './../outlets/TodoFooterOutlet';
 import TodoHeader from './TodoHeader';
 import TodoSearch from './TodoSearch';
 import Credits from './Credits';
 import { Todo } from './../todoProcesses';
 
 import * as css from './styles/todoApp.m.css';
+import Outlet from '@dojo/framework/routing/Outlet';
+import { MatchDetails } from '@dojo/framework/routing/interfaces';
+import TodoList from './TodoList';
+import TodoFooter from './TodoFooter';
 
 export interface TodoAppProperties {
 	todos: Todo[];
@@ -61,8 +63,22 @@ export default class TodoApp extends ThemedMixin(WidgetBase)<TodoAppProperties> 
 					setCurrentTodo
 				}),
 				todoCount > 0 ? w(TodoSearch, { search, searchValue }) : null,
-				todoCount > 0 ? w(TodoListOutlet, { todos, searchValue, toggleTodo, removeTodo, editTodo }) : null,
-				todoCount > 0 ? w(TodoFooterOutlet, { activeCount, todoCount, clearCompleted }) : null
+				todoCount > 0 ? w(Outlet, { id: 'view', renderer: ({ router, params, queryParams }: MatchDetails) => {
+					return w(TodoList, {
+						todos,
+						searchValue,
+						toggleTodo,
+						removeTodo,
+						view: params.view,
+						filter: queryParams.filter,
+						editTodo({ id }: {id: string}) {
+							const link = router.link('edit', { id });
+							link && router.setPath(link);
+						}});
+				}}) : null,
+				todoCount > 0 ? w(Outlet, { id: 'view', renderer: ({ params, queryParams }: MatchDetails) => {
+					return w(TodoFooter, { view: params.view, filter: queryParams.filter, activeCount, todoCount, clearCompleted });
+				}}) : null
 			]),
 			w(Credits, {})
 		]);
